--- .dockerignore
@@ -1,2 +1,0 @@
-**/build
-**/__pycache__
