+++ .gitignore
@@ -3,4 +3,3 @@
 **/.DS_Store
 **/.vscode
 *.ipynb
-**/build
