--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "external/pybind11"]
-	path = external/pybind11
-	url = https://github.com/pybind/pybind11
