+++ Dockerfile
@@ -12,8 +12,6 @@ COPY . .
 
 RUN pip install -r requirements.txt --no-cache-dir
 
-RUN make
-
 CMD [ "uvicorn", "app:app", "--host", "0.0.0.0", "--port", "7000", "--reload" ]
 
 EXPOSE 7000
