+++ app.py
@@ -4,6 +4,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import ORJSONResponse
 from tortoise.contrib.fastapi import register_tortoise
 
+import dotenv
 import os
 
 import common.config as config
