+++ common/config.py
@@ -12,6 +12,5 @@ except KeyError:
 
 if BRANCH is None:
     GOOGLE_REDIRECT_URI = "http://localhost:7000/v1/auth/google/callback"
-    os.system("make")
 else:
     GOOGLE_REDIRECT_URI = f"https://dene.sh{PREFIX}/v1/auth/google/callback"
