--- common/logger.py
@@ -1,7 +1,0 @@
-from models import Log
-
-async def log(message: str):
-    log = Log(
-        text=message,
-    )
-    await log.save()
