--- common/names.json
@@ -1,328 +1,0 @@
-{
-    "first": {
-        "male": [
-            "John",
-            "Michael",
-            "William",
-            "David",
-            "James",
-            "Robert",
-            "Joseph",
-            "Daniel",
-            "Thomas",
-            "Matthew",
-            "Christopher",
-            "Andrew",
-            "Joshua",
-            "Nicholas",
-            "Anthony",
-            "Ryan",
-            "Kevin",
-            "Brian",
-            "Jason",
-            "Jeffrey",
-            "Timothy",
-            "Steven",
-            "Mark",
-            "Charles",
-            "Scott",
-            "Eric",
-            "Stephen",
-            "Patrick",
-            "Edward",
-            "Jonathan",
-            "Benjamin",
-            "Adam",
-            "Gregory",
-            "Jeremy",
-            "Paul",
-            "Kenneth",
-            "Alexander",
-            "Dennis",
-            "Tyler",
-            "Brandon",
-            "Samuel",
-            "Travis",
-            "Jesse",
-            "Bryan",
-            "Peter",
-            "Cody",
-            "Shawn",
-            "Jared",
-            "Juan",
-            "Justin",
-            "Trevor",
-            "Kyle",
-            "Keith",
-            "Roger",
-            "Raymond",
-            "Aaron",
-            "Christian",
-            "Jose",
-            "Ethan",
-            "Derek",
-            "Adam",
-            "Nathan",
-            "Wesley",
-            "Marcus",
-            "Craig",
-            "Dustin",
-            "Antonio",
-            "Caleb",
-            "Philip",
-            "Russell",
-            "Brent",
-            "Corey",
-            "Luis",
-            "Bradley",
-            "Shane",
-            "Adrian",
-            "Gabriel",
-            "Miguel",
-            "Oscar",
-            "Evan",
-            "Casey",
-            "Luke",
-            "Mitchell",
-            "Cameron",
-            "Blake",
-            "Spencer",
-            "Lucas",
-            "Keith",
-            "Dylan",
-            "Jesus",
-            "Garrett",
-            "Jack",
-            "Brett",
-            "Ian",
-            "Victor",
-            "Chad",
-            "Curtis",
-            "Randy",
-            "Mario",
-            "Joel"
-        ],
-        "female": [
-            "Emma",
-            "Olivia",
-            "Ava",
-            "Isabella",
-            "Sophia",
-            "Mia",
-            "Amelia",
-            "Harper",
-            "Evelyn",
-            "Abigail",
-            "Emily",
-            "Ella",
-            "Scarlett",
-            "Grace",
-            "Chloe",
-            "Lily",
-            "Aria",
-            "Zoe",
-            "Charlotte",
-            "Hannah",
-            "Avery",
-            "Addison",
-            "Eleanor",
-            "Natalie",
-            "Luna",
-            "Penelope",
-            "Aubrey",
-            "Stella",
-            "Mila",
-            "Leah",
-            "Hazel",
-            "Violet",
-            "Aurora",
-            "Savannah",
-            "Audrey",
-            "Bella",
-            "Claire",
-            "Skylar",
-            "Lucy",
-            "Paisley",
-            "Everly",
-            "Anna",
-            "Caroline",
-            "Nova",
-            "Genesis",
-            "Emilia",
-            "Kennedy",
-            "Sofia",
-            "Madison",
-            "Elena",
-            "Ruby",
-            "Aaliyah",
-            "Eva",
-            "Katherine",
-            "Alice",
-            "Josephine",
-            "Madelyn",
-            "Cora",
-            "Riley",
-            "Eliana",
-            "Willow",
-            "Layla",
-            "Nora",
-            "Scarlet",
-            "Ariana",
-            "Maya",
-            "Eleanor",
-            "Hannah",
-            "Piper",
-            "Grace",
-            "Zara",
-            "Ellie",
-            "Aurora",
-            "Aubrey",
-            "Zoey",
-            "Harlow",
-            "Victoria",
-            "Natalie",
-            "Lily",
-            "Nova",
-            "Hazel",
-            "Emilia",
-            "Stella",
-            "Isla",
-            "Chloe",
-            "Quinn",
-            "Luna",
-            "Sofia",
-            "Mila",
-            "Lucy",
-            "Ella",
-            "Aria"
-        ]
-    },
-    "last": [
-        "Anderson",
-        "Baker",
-        "Carter",
-        "Davis",
-        "Edwards",
-        "Fisher",
-        "Garcia",
-        "Hill",
-        "Irwin",
-        "Jones",
-        "Keller",
-        "Lopez",
-        "Miller",
-        "Nelson",
-        "Owens",
-        "Perez",
-        "Quinn",
-        "Ramirez",
-        "Smith",
-        "Taylor",
-        "Underwood",
-        "Vasquez",
-        "Williams",
-        "Xavier",
-        "Young",
-        "Zimmerman",
-        "Adams",
-        "Brown",
-        "Cooper",
-        "Diaz",
-        "Evans",
-        "Fletcher",
-        "Gomez",
-        "Harrison",
-        "Ingram",
-        "Johnson",
-        "Kingsley",
-        "Lee",
-        "Morgan",
-        "Nixon",
-        "Oliver",
-        "Parker",
-        "Reyes",
-        "Sanders",
-        "Thomas",
-        "Underhill",
-        "Vega",
-        "Wright",
-        "Xander",
-        "Yates",
-        "Zhang",
-        "Armstrong",
-        "Barnes",
-        "Clark",
-        "Duncan",
-        "Elliott",
-        "Freeman",
-        "Grant",
-        "Hartman",
-        "Inglewood",
-        "Jefferson",
-        "Kennedy",
-        "Lawson",
-        "Morton",
-        "Nash",
-        "Olsen",
-        "Peterson",
-        "Quigley",
-        "Riley",
-        "Stewart",
-        "Turner",
-        "Vanderbilt",
-        "Watson",
-        "Xavier",
-        "York",
-        "Zane",
-        "Aldridge",
-        "Bennett",
-        "Cunningham",
-        "Drake",
-        "Eaton",
-        "Ferguson",
-        "Gordon",
-        "Henderson",
-        "Isaacs",
-        "Jenkins",
-        "Kim",
-        "Larson",
-        "Matthews",
-        "Newton",
-        "O'Donnell",
-        "Porter",
-        "Quill",
-        "Reynolds",
-        "Sullivan",
-        "Tucker",
-        "Vance",
-        "Wallace",
-        "Xiong",
-        "Yoder",
-        "Zimmer",
-        "Ayers",
-        "Browning",
-        "Conway",
-        "Davenport",
-        "Ellis",
-        "Fleming",
-        "Griffith",
-        "Harrison",
-        "Ingram",
-        "Jensen",
-        "Keane",
-        "Lyons",
-        "Montgomery",
-        "Noble",
-        "Osborne",
-        "Pennington",
-        "Quinn",
-        "Randall",
-        "Stevens",
-        "Thornton",
-        "Vaughn",
-        "Wells",
-        "Xu",
-        "Youngblood",
-        "Ziegler"
-    ]
-}
