+++ requirements.txt
@@ -8,5 +8,3 @@ python-jose[cryptography]==3.3.0
 passlib[argon2]==1.7.4
 httpx==0.25.0
 aerich==0.7.2
-numpy==1.26.0
-matplotlib==3.5.1
