+++ routers/accounts.py
@@ -1,16 +1,24 @@
 from fastapi import APIRouter, HTTPException, status, Depends, Response
 
-from models import Account, AccountResponse
+from models import Account
 
 from typing import Annotated, Optional
 
 from dependencies.auth import require_account, require_staff_token, Token
 from common.auth import hash_password, check_email
-from common.logger import log
 
 from pydantic import BaseModel
 
 router = APIRouter(tags=["accounts"])
+
+
+class AccountResponse(BaseModel):
+    id: int
+    email: Optional[str]
+    google_id: Optional[str]
+    created_at: int
+    updated_at: int
+    type: str
 
 
 @router.get("")
@@ -19,8 +27,14 @@ async def get_accounts(
 ):
     accounts = await Account.all()
     return [
-        await AccountResponse.create(account)
-        for account in accounts
+        AccountResponse(
+            id=account.id,
+            email=account.email,
+            google_id=account.google_id,
+            created_at=account.created_at,
+            updated_at=account.updated_at,
+            type=account.type,
+        ) for account in accounts
     ]
 
 
@@ -28,7 +42,14 @@ async def get_accounts(
 async def get_me(
     account: Annotated[Account, Depends(require_account)],
 ):
-    return await AccountResponse.create(account)
+    return AccountResponse(
+        id=account.id,
+        email=account.email,
+        google_id=account.google_id,
+        created_at=account.created_at,
+        updated_at=account.updated_at,
+        type=account.type,
+    )
 
 
 @router.get("/{id}")
@@ -41,7 +62,13 @@ async def get_account(
     if account is None:
         raise HTTPException(status.HTTP_404_NOT_FOUND)
 
-    return await AccountResponse.create(account)
+    return AccountResponse(
+        email=account.email,
+        google_id=account.google_id,
+        created_at=account.created_at,
+        updated_at=account.updated_at,
+        type=account.type,
+    )
 
 
 @router.post("/me/change-password")
@@ -51,7 +78,6 @@ async def change_password(
 ):
     account.password = hash_password(password)
     await account.update()
-    await log(f"Password changed for account {str(account.id)} ({account.first_name} {account.last_name}).")
 
 
 @router.post("/me/change-email")
@@ -66,7 +92,6 @@ async def change_email(
 
     account.email = email
     await account.update()
-    await log(f"Email changed for account {str(account.id)} ({account.first_name} {account.last_name}).")
 
 
 @router.delete("/me")
@@ -75,7 +100,6 @@ async def delete_me(
 ):
     response = Response()
     response.delete_cookie("token")
-    await log(f"Account deleted: {account.id} ({account.first_name} {account.last_name})")
     await account.delete()
     return response
 
@@ -90,5 +114,4 @@ async def delete_account(
     if account is None:
         raise HTTPException(status.HTTP_404_NOT_FOUND)
 
-    await log(f"Account deleted: {account.id} ({account.first_name} {account.last_name})")
     await account.delete()
